chore: refactor Netlify Graph logic into a reusable package with types and docstrings
