+++ package.json
@@ -146,6 +146,7 @@
     "multiparty": "^4.2.1",
     "netlify": "^10.1.1",
     "netlify-headers-parser": "^6.0.0",
+    "netlify-onegraph-internal": "0.0.2",
     "netlify-redirect-parser": "^13.0.0",
     "netlify-redirector": "^0.2.1",
     "node-fetch": "^2.6.0",
