+++ tests/assets/netlifyGraphOperationsLibrary.graphql
@@ -1,4 +1,4 @@
-query ExampleQuery($package: String!) @netligraph(id: "d86699fb-ddfc-4833-9d9a-f3497cb7c992", doc: "A test query to snapshot") {
+query ExampleQuery($package: String!) @netlify(id: "d86699fb-ddfc-4833-9d9a-f3497cb7c992", doc: "A test query to snapshot") {
   npm {
     package(name: $package) {
       id
